feat: migrate landing page functionality to development page and add WebContainer model
