+++ src/app/modules/dashboard/models/analysisResult.model.ts
@@ -3,13 +3,14 @@ import { BrandIdentityModel } from './brand-identity.model';
 import { DiagramModel } from './diagram.model';
 import { LandingModel } from './landing.model';
 import { BusinessPlanModel } from './businessPlan.model';
+import { WebContainerModel } from './webcontainer.model';
 
 export interface AnalysisResultModel {
   id?: string;
   architectures: ArchitectureModel[];
   businessPlan?: BusinessPlanModel;
   design: DiagramModel;
-  development: string;
+  development: WebContainerModel;
   branding: BrandIdentityModel;
   landing: LandingModel;
   testing: string;
