--- src/app/modules/dashboard/pages/show-landing/show-landing.css
@@ -1,3 +1,0 @@
-:host {
-  display: block;
-}
