--- src/app/modules/dashboard/pages/show-landing/show-landing.html
@@ -1,183 +1,0 @@
-<div class="bg-gray-900 p-8 rounded-xl">
-  @if (isLoaded() === true) {
-  <div class="h-screen w-full flex items-center justify-center">
-    <app-loader />
-  </div>
-  }
-  @if (!project.analysisResultModel.landing) {
-  <section class="mb-12">
-    <h1 class="text-white text-4xl font-bold mb-3">Generate Project Landing</h1>
-    <p class="text-gray-400 text-lg max-w-2xl">
-      Select a frontend technology to generate a high-converting landing page
-      for your project.
-    </p>
-  </section>
-
-  <section class="mb-12">
-    <h2
-      class="text-white text-2xl font-semibold mb-6 border-b border-gray-800 pb-3"
-    >
-      Choose Your Stack
-    </h2>
-
-    <div class="grid md:grid-cols-3 gap-3">
-      @for (stack of stacks; track stack.id) {
-      <div
-        class="bg-gray-800 p-5 rounded-lg border-2 transition-colors cursor-pointer hover:border-2 hover:border-primary"
-        [class.border-primary]="selectedStackId === stack.id"
-        [class.border-transparent]="selectedStackId !== stack.id"
-        (click)="selectStack(stack.id)"
-      >
-        <div class="flex items-start">
-          <!-- Icon -->
-          <div
-            class="w-8 h-8 rounded flex items-center justify-center mr-3 mt-0.5"
-            [style.backgroundColor]="stack.color + '10'"
-          >
-            <span class="text-lg" [style.color]="stack.color">
-              {{ stack.icon }}
-            </span>
-          </div>
-
-          <!-- Content -->
-          <div class="flex-1">
-            <div class="flex justify-between items-start">
-              <h3 class="text-white font-medium">{{ stack.name }}</h3>
-
-              <!-- Visual checkbox -->
-              <div
-                class="w-5 h-5 rounded-full border-2 flex items-center justify-center"
-                [class.border-primary]="selectedStackId === stack.id"
-                [class.border-gray-600]="selectedStackId !== stack.id"
-              >
-                @if (selectedStackId === stack.id) {
-                <div class="w-3 h-3 rounded-full bg-primary"></div>
-                }
-              </div>
-            </div>
-
-            <p class="text-gray-400 text-sm mt-1">{{ stack.description }}</p>
-
-            <!-- Badges -->
-            @if (stack.badges) {
-            <div class="flex flex-wrap gap-2 mt-3">
-              @for (badge of stack.badges; track badge) {
-              <span
-                class="text-xs px-2 py-1 rounded-full bg-gray-700 text-gray-200"
-              >
-                {{ badge }}
-              </span>
-              }
-            </div>
-            }
-          </div>
-        </div>
-      </div>
-      }
-    </div>
-  </section>
-
-  <section class="mb-8">
-    <h2 class="text-white text-2xl font-semibold mb-4">
-      Advance Configurations
-    </h2>
-
-    <div class="flex flex-wrap gap-3">
-      @for (option of pageOptions; track option.id) {
-      <div class="relative">
-        <!-- Compact card - Main visible element -->
-        <div
-          class="flex items-center p-2 pl-1 pr-3 rounded-lg border cursor-pointer transition-all duration-200 group"
-          [class.border-primary]="option.enabled"
-          [class.border-gray-700]="!option.enabled"
-          [class.bg-gray-800]="option.enabled"
-          [class.transform]="hoveredOption === option.id"
-          [class.-translate-y-2]="hoveredOption === option.id"
-          (mouseenter)="hoveredOption = option.id"
-          (mouseleave)="hoveredOption = null"
-          (click)="toggleOption(option.id)"
-        >
-          <!-- Animated checkbox -->
-          <div
-            class="w-5 h-5 rounded-md border mr-2 flex-shrink-0 flex items-center justify-center transition-colors"
-            [class.border-primary]="option.enabled"
-            [class.border-gray-500]="!option.enabled"
-            [class.bg-primary]="option.enabled"
-          >
-            @if (option.enabled) {
-            <span class="text-white text-xs">✓</span>
-            }
-          </div>
-
-          <!-- Icon + Name -->
-          <div class="flex items-center">
-            <span class="text-sm mr-2">{{ option.icon }}</span>
-            <span class="text-sm">{{ option.name }}</span>
-          </div>
-        </div>
-
-        <!-- Detailed card - Pop-in effect -->
-        @if (hoveredOption === option.id) {
-        <div
-          class="absolute z-20 bottom-full left-0 mb-3 w-72 p-4 bg-gray-800 border-2 border-primary rounded-lg shadow-2xl animate-popIn"
-        >
-          <div class="flex items-start">
-            <span class="text-2xl mr-3 mt-1">{{ option.icon }}</span>
-            <div>
-              <h3 class="font-bold text-lg text-white mb-1">
-                {{ option.name }}
-              </h3>
-              <p class="text-gray-400 text-sm mb-3">{{ option.description }}</p>
-
-              @if (option.features) {
-              <div class="flex flex-wrap gap-2">
-                @for (feature of option.features; track feature) {
-                <span
-                  class="text-xs px-3 py-1 rounded-full bg-primary/10 text-gray-200 border border-primary/30"
-                >
-                  {{ feature }}
-                </span>
-                }
-              </div>
-              }
-            </div>
-          </div>
-
-          <!-- Visual indicator arrow -->
-          <div
-            class="absolute -bottom-2 left-4 w-4 h-4 bg-gray-800 border-r-2 border-b-2 border-primary transform rotate-45"
-          ></div>
-        </div>
-        }
-      </div>
-      }
-    </div>
-  </section>
-
-  <button
-    class="w-full bg-primary hover:bg-primary/90 text-white py-3 rounded-lg font-medium transition-colors"
-    [disabled]="!selectedStackId"
-    (click)="onGenerateLanding()"
-  >
-    Generate Landing Page
-  </button>
-  } @else {
-  <div class="text-center py-12">
-    <div class="text-6xl mb-4">🎉</div>
-    <h1 class="text-white text-2xl font-bold mb-3">Landing Page Already Generated!</h1>
-    <p class="text-gray-400 text-lg mb-6">
-      Your landing page has been successfully generated with your selected options.
-    </p>
-    <p class="text-gray-400">
-      Stack: <span class="text-primary">{{ project.analysisResultModel.landing.selectedOptions.stack }}</span>
-    <button
-      class="w-full bg-primary hover:bg-primary/90 text-white py-3 rounded-lg font-medium transition-colors"
-      (click)="redirectToReactApp(this.id)"
-    >
-      Go to Landing Editor
-    </button>
-
-    </p>
-  </div>
-  }
-</div>
